AB#131 done - removed duplicate api calls
